com.unity.ai.generators@1.5.0-pre.1 ## [1.5.0-pre.1] - 2026-01-05
### Added

- Generation workflows integrated into AI Assistant
- Added spritesheet generation
- Added animation stationary and loop modes
- Added skybox material creation from cubemap
- Added audio clip editing tool for trimming and modifying sounds

+++ Modules/Unity.AI.Generators.Asset/AssetReferenceDeletionProcessor.cs
@@ -1,5 +1,8 @@
-﻿using System;
+#define AI_TOOLKIT_GENERATION_CLEANUP
+#if AI_TOOLKIT_GENERATION_CLEANUP
+using System;
 using System.IO;
+using Unity.AI.Toolkit.Asset;
 using UnityEditor;
 using UnityEngine;
 
@@ -11,25 +14,28 @@ namespace Unity.AI.Generators.Asset
         {
             var guid = AssetDatabase.AssetPathToGUID(assetPath);
             var asset = new AssetReference { guid = guid };
-            var folderPath = asset.GetGeneratedAssetsPath();
-
-            if (!Directory.Exists(folderPath))
-                return AssetDeleteResult.DidNotDelete;
-
-            try
+            if (asset.IsValid())
             {
-                var deletedFolderPath = folderPath + "_deleted";
-                if (!Directory.Exists(deletedFolderPath))
-                    Directory.Move(folderPath, deletedFolderPath);
-                else
+                var folderPath = asset.GetGeneratedAssetsPath();
+
+                if (!Directory.Exists(folderPath))
+                    return AssetDeleteResult.DidNotDelete;
+
+                try
                 {
-                    CopyContents(folderPath, deletedFolderPath);
-                    Directory.Delete(folderPath, true);
+                    var deletedFolderPath = folderPath + "_deleted";
+                    if (!Directory.Exists(deletedFolderPath))
+                        Directory.Move(folderPath, deletedFolderPath);
+                    else
+                    {
+                        CopyContents(folderPath, deletedFolderPath);
+                        Directory.Delete(folderPath, true);
+                    }
                 }
-            }
-            catch
-            {
-                Debug.Log($"Some generated assets remain in '{folderPath}'.");
+                catch
+                {
+                    Debug.Log($"Some generated assets remain in '{folderPath}'.");
+                }
             }
 
             return AssetDeleteResult.DidNotDelete;
@@ -51,3 +57,4 @@ namespace Unity.AI.Generators.Asset
         }
     }
 }
+#endif
